fix: Fusión del componente de garajes disponibles con el listado de garajes de administrador
