--- src/app/components/rental/available-garages/available-garages.component.html
@@ -1,31 +1,0 @@
-<ion-grid>
-  <ion-row>
-    <ion-col size="12" size-md="4">
-      <ion-input [(ngModel)]="nameFilter" placeholder="Filtrar por nombre"></ion-input>
-    </ion-col>
-    <ion-col size="12" size-md="4">
-      <ion-input [(ngModel)]="priceFilter" placeholder="Filtrar por precio" type="number"></ion-input>
-    </ion-col>
-    <ion-col size="12" size-md="4">
-      <ion-input [(ngModel)]="dimensionFilter" placeholder="Filtrar por dimensiones" type="number"></ion-input>
-    </ion-col>
-  </ion-row>
-
-  <ion-row *ngFor="let garage of filteredGarages">
-    <ion-col size="12">
-      <ion-card>
-        <ion-card-header>
-          <ion-card-title><a [routerLink]="['/aparKing/garages', garage.id]">{{ garage.name }}</a></ion-card-title>
-          <ion-card-subtitle>{{ garage.description }}</ion-card-subtitle>
-        </ion-card-header>
-        <ion-card-content>
-          <p>Dimensiones: {{ garage.height }} x {{ garage.width }} x {{ garage.length }}</p>
-          <p>Precio: {{ garage.price }}</p>
-          <p>Día de subida: {{ garage.creation_date }}</p>
-          <p>Estado: {{ garage.is_active ? 'Active' : 'Inactive' }}</p>
-          <p>Propietario: {{ garage.owner }}</p>
-        </ion-card-content>
-      </ion-card>
-    </ion-col>
-  </ion-row>
-</ion-grid>
