--- src/app/components/rental/available-garages/available-garages.component.scss
@@ -1,11 +1,0 @@
-.garage-list {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  padding: 16px;
-}
-
-.garage-card {
-  max-width: 400px;
-  margin: 16px;
-}
