+++ src/app/components/rental/garage-list/garage-list.component.html
@@ -1,76 +1,91 @@
-<ion-content class="ion-padding">
-  <div class="ion-text-center">
-    <ion-card>
-      <ion-card-header>
-        <ion-toolbar>
-          <ion-label>Listado de alquileres de garajes</ion-label>
-          <ion-button
-            routerLink="/garage/create"
-            size="small"
-            slot="end"
-            shape="round"
-            color="primary"
-            >Crear alquiler</ion-button
+<ion-grid>
+  <ion-row>
+    <ion-col size="12" size-md="4">
+      <ion-input
+        [(ngModel)]="nameFilter"
+        placeholder="Filtrar por nombre"
+      ></ion-input>
+    </ion-col>
+    <ion-col size="12" size-md="4">
+      <ion-input
+        [(ngModel)]="priceFilter"
+        placeholder="Filtrar por precio"
+        type="number"
+        min="0"
+        max="9999"
+      ></ion-input>
+    </ion-col>
+    <ion-col size="12" size-md="4">
+      <ion-input
+        [(ngModel)]="dimensionFilter"
+        placeholder="Filtrar por dimensiones"
+        type="number"
+        min="0"
+        max="9999"
+      ></ion-input>
+    </ion-col>
+  </ion-row>
+
+  <ion-row *ngFor="let garage of filteredGarages">
+    <ion-col size="12">
+      <ion-card>
+        <ion-card-header>
+          <ion-card-title
+            ><a [routerLink]="['/aparKing/garages', garage.id]">{{
+              garage.name
+            }}</a></ion-card-title
           >
-        </ion-toolbar>
-      </ion-card-header>
+          <ion-card-subtitle>{{ garage.description }}</ion-card-subtitle>
+        </ion-card-header>
+        <ion-card-content>
+          <p>
+            Dimensiones: {{ garage.height }} x {{ garage.width }} x
+            {{ garage.length }}
+          </p>
+          <p>Precio: {{ garage.price }}</p>
+          <p>Día de subida: {{ garage.creation_date }}</p>
+          <p>Estado: {{ garage.is_active ? "Activo" : "Inactivo" }}</p>
+          <p>Propietario: {{ garage.owner }}</p>
+        </ion-card-content>
+      </ion-card>
+    </ion-col>
+  </ion-row>
 
-      <ion-grid class="ion-text-center">
-        <ion-row class="header-row first-row">
-          <ion-col>Id</ion-col>
-          <ion-col>Nombre</ion-col>
-          <ion-col>Descripción</ion-col>
-          <ion-col>Altura</ion-col>
-          <ion-col>Anchura</ion-col>
-          <ion-col>Longitud</ion-col>
-          <ion-col>Precio</ion-col>
-          <ion-col>Fecha de creación</ion-col>
-          <ion-col>Activo</ion-col>
-          <ion-col>Propietario</ion-col>
-          <ion-col>Acciones</ion-col>
-        </ion-row>
-
-        <ng-container *ngFor="let garage of garages">
-          <ion-row>
-            <ion-col>{{ garage.id }}</ion-col>
-            <ion-col
-              ><a href="aparKing/garages/{{ garage.id }}">{{
-                garage.name
-              }}</a></ion-col
-            >
-            <ion-col>{{ garage.description }}</ion-col>
-            <ion-col>{{ garage.height }}</ion-col>
-            <ion-col>{{ garage.width }}</ion-col>
-            <ion-col>{{ garage.length }}</ion-col>
-            <ion-col>{{ garage.price }}</ion-col>
-            <ion-col>{{ garage.creation_date }}</ion-col>
-            <ion-col>{{ garage.is_active }}</ion-col>
-            <ion-col>{{ garage.owner }}</ion-col>
-            <ion-col>
-              <ion-button
-                [routerLink]="'aparKing/garages/' + garage.id"
-                color="secondary"
-                shape="round"
-                size="small"
-                >Editar</ion-button
-              >
-              <ion-button
-                [routerLink]="'aparKing/garages/' + garage.id"
-                color="danger"
-                shape="round"
-                size="small"
-                >Eliminar</ion-button
-              >
-            </ion-col>
-          </ion-row>
-        </ng-container>
-        <ion-infinite-scroll>
-          <ion-infinite-scroll-content
-            loadingText="Por favor espere..."
-            loadingSpinner="bubbles"
-          ></ion-infinite-scroll-content>
-        </ion-infinite-scroll>
-      </ion-grid>
-    </ion-card>
-  </div>
-</ion-content>
+  <ion-row *ngFor="let garage of garages">
+    <ion-col size="12">
+      <ion-card>
+        <ion-card-header>
+          <ion-card-title
+            ><a [routerLink]="['/aparKing/garages', garage.id]">{{
+              garage.name
+            }}</a></ion-card-title
+          >
+          <ion-card-subtitle>{{ garage.description }}</ion-card-subtitle>
+        </ion-card-header>
+        <ion-card-content>
+          <p>
+            Dimensiones: {{ garage.height }} x {{ garage.width }} x
+            {{ garage.length }}
+          </p>
+          <p>Precio: {{ garage.price }}</p>
+          <p>Día de subida: {{ garage.creation_date }}</p>
+          <p>Estado: {{ garage.is_active ? "Activo" : "Inactivo" }}</p>
+          <p>Propietario: {{ garage.owner }}</p>
+          <ion-button
+            [routerLink]="garage.id"
+            color="secondary"
+            shape="round"
+            size="small"
+            >Ver detalles</ion-button
+          >
+        </ion-card-content>
+      </ion-card>
+    </ion-col>
+  </ion-row>
+  <ion-infinite-scroll>
+    <ion-infinite-scroll-content
+      loadingText="Por favor espere..."
+      loadingSpinner="bubbles"
+    ></ion-infinite-scroll-content>
+  </ion-infinite-scroll>
+</ion-grid>
