+++ src/app/components/rental/garage-list/garage-list.component.scss
@@ -69,3 +69,16 @@
     text-align: center;
     size: 6px;
   }
+
+  .garage-list {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-around;
+    padding: 16px;
+  }
+  
+  .garage-card {
+    max-width: 400px;
+    margin: 16px;
+  }
+
