+++ src/app/tab2/tab2-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { Tab2Page } from './tab2.page';
 import { GarageDetailComponent } from '../components/rental/garage-detail/garage-detail.component';
+import { GarageListComponent } from '../components/rental/garage-list/garage-list.component';
 
 const routes: Routes = [
   {
