+++ src/app/tabs/tabs.module.ts
@@ -9,7 +9,6 @@ import { TabsPageRoutingModule } from './tabs-routing.module';
 import { RouterModule } from '@angular/router';
 import { GarageDetailComponent } from '../components/rental/garage-detail/garage-detail.component';
 import { GarageListComponent } from '../components/rental/garage-list/garage-list.component';
-import { AvailableGaragesComponent } from '../components/rental/available-garages/available-garages.component';
 import { MyGaragesComponent } from '../components/rental/my-garages/my-garages.component';
 import { RestService } from '../services/rest.service';
 import { TabsPage } from './tabs.page';
@@ -23,8 +22,8 @@ import { TabsPage } from './tabs.page';
     HttpClientModule,
     RouterModule,
   ],
-  declarations: [TabsPage, GarageListComponent, GarageDetailComponent, AvailableGaragesComponent, MyGaragesComponent],
-  exports: [GarageListComponent, GarageDetailComponent, AvailableGaragesComponent, MyGaragesComponent],
+  declarations: [TabsPage, GarageListComponent, GarageDetailComponent, MyGaragesComponent],
+  exports: [GarageListComponent, GarageDetailComponent, MyGaragesComponent],
   providers: [RestService],
 })
 export class TabsPageModule {}
